// uldb/ngx-unity/src/app/unity-setup/unity-setup-integration/usi-network-controllers/usinc-cisco-meraki/usinc-cisco-meraki.type.ts
import { DeviceProxy } from "src/app/shared/SharedEntityTypes/device-common-utils.type";
import { DeviceMonitoringType } from "src/app/shared/SharedEntityTypes/devices-monitoring.type";
import { UnityScheduleType } from "src/app/shared/SharedEntityTypes/schedule.type";

export interface MerakiAccountType {
    uuid: string;
    name: string;
    account_url: string;
    port?: string;
    collector: string;
    username: string;
    password: string;
    schedule_meta: UnityScheduleType;
    status: string;
    proxy: DeviceProxy;
    tags: string[];
    monitoring: DeviceMonitoringType;
}